+++ inc/cropper/src/views/image-edit.js
@@ -1,3 +1,4 @@
+import { applyFilters } from '@wordpress/hooks';
 import Media from '@wordpress/media';
 import template from '@wordpress/template';
 import ImageEditSizes from './image-edit-sizes';
@@ -15,11 +16,42 @@ const ImageEditView = Media.View.extend( {
 			this.model.set( { size: 'full' } );
 		}
 
+		// Get current size from block attributes if available.
+		this.setSizeFromBlock();
+
 		// Re-render on certain updates.
 		this.listenTo( this.model, 'change:url', this.onUpdate );
 
 		// Initial render.
 		this.onUpdate();
+	},
+	setSizeFromBlock() {
+		if ( ! wp || ! wp.data ) {
+			return;
+		}
+
+		const selectedBlock = wp.data.select( 'core/block-editor' ).getSelectedBlock();
+		if ( ! selectedBlock ) {
+			return;
+		}
+
+		const sizeForBlock = applyFilters(
+			`smartmedia.cropper.selectSizeFromBlockAttributes.${ selectedBlock.name.replace( /\W+/g, '.' ) }`,
+			null,
+			selectedBlock
+		);
+
+		const size = applyFilters(
+			'smartmedia.cropper.selectSizeFromBlockAttributes',
+			sizeForBlock,
+			selectedBlock
+		);
+
+		if ( ! size ) {
+			return;
+		}
+
+		this.model.set( { size } );
 	},
 	onUpdate() {
 		const views = [];